refactor(fields): единый BaseCandidatePicker — 3 пикера базы → один (issue #73, шаг 2)
Три параллельных пикера базового экземпляра (CatalogBaseEntryPicker в каталоге,
BaseCandidatePicker в редакторе документов, локальный BaseEntryPickerModal в системном
каталоге) сведены к одному презентационному компоненту в document-sets/fields/
BaseCandidatePicker.tsx (+ общие хелперы BaseCandidate/SCOPE_TIER/ancestorTypeIds/
parseBaseRef). Различались только источником кандидатов — теперь это ответственность
вызывающего: он строит BaseCandidate[] и решает формат хранения _baseRef (документ —
{kind,id}; каталог/система — голый id). Поведение сохранено.

Part of #73.
