+++ src/client/src/features/document-sets/catalog/index.tsx
@@ -25,6 +25,7 @@ import { EntryDataSetBindings } from './EntryDataSetBindings';
 import {
   SCOPE_COLORS, ComplexFieldGroup, ArrayFieldEditor, DocRefCatalogPickerField,
   PrimitiveInput, FileField, ImageField,
+  BaseCandidatePicker, SCOPE_TIER, type BaseCandidate,
 } from '../fields';
 
 export function ScopedCatalogPanel({ scope, scopeId, allDocTypes, setId }: {
@@ -159,54 +160,8 @@ export function ScopedCatalogPanel({ scope, scopeId, allDocTypes, setId }: {
   );
 }
 
-// ─── Base entry picker for catalog entry form ─────────────────────────────────
-
-function CatalogBaseEntryPicker({ open, onOpenChange, parentType, setId, scope, scopeId, onSelect }: {
-  open: boolean; onOpenChange: (o: boolean) => void;
-  parentType: DocumentType; setId?: string;
-  scope: CatalogScope; scopeId: string | null;
-  onSelect: (entry: CommonDataEntry) => void;
-}) {
-  const [search, setSearch] = useState('');
-  const { data: setCatalogEntries = [] } = useCommonDataForSet({
-    setId: setId ?? '', typeId: parentType.id, enabled: open && !!setId,
-  });
-  const { data: scopeEntries = [] } = useListCommonData({
-    scope, scopeId: scopeId ?? undefined, typeId: parentType.id,
-    enabled: open && !setId && scope !== 'System',
-  });
-  const { data: systemEntries = [] } = useListCommonData({
-    scope: 'System', typeId: parentType.id, enabled: open,
-  });
-  const allEntries: CommonDataEntry[] = setId
-    ? setCatalogEntries
-    : [...scopeEntries, ...systemEntries.filter(e => !scopeEntries.some(s => s.id === e.id))];
-  const filtered = allEntries.filter(e => e.displayName.toLowerCase().includes(search.toLowerCase()));
-
-  return (
-    <Modal open={open} onOpenChange={onOpenChange} title={`Базовый экземпляр: ${parentType.name}`}>
-      <div className="space-y-4">
-        <input value={search} onChange={e => setSearch(e.target.value)} placeholder="Поиск..." autoFocus
-          className="w-full border border-stroke-strong rounded-md px-3 py-2 text-sm focus:outline-none focus-visible:ring-2 focus-visible:ring-brand bg-surface" />
-        {filtered.length === 0 ? (
-          <p className="text-sm text-fg4 text-center py-4">
-            Нет записей типа «{parentType.name}».
-          </p>
-        ) : (
-          <div className="space-y-1 max-h-72 overflow-y-auto">
-            {filtered.map(entry => (
-              <button key={entry.id} type="button" onClick={() => { onSelect(entry); onOpenChange(false); }}
-                className="w-full flex items-center gap-3 px-3 py-2 text-sm text-left rounded-md hover:bg-brand-subtle transition-colors">
-                <Link2 size={13} className="text-brand shrink-0" />
-                <span className="flex-1 font-medium text-fg1 truncate">{entry.displayName}</span>
-              </button>
-            ))}
-          </div>
-        )}
-      </div>
-    </Modal>
-  );
-}
+// Базовый экземпляр каталога использует общий BaseCandidatePicker (issue #73, шаг 2) —
+// кандидаты (записи родительского типа по скопам) строятся ниже из parentEntries.
 
 // ─── Catalog entry form (create + edit, shared by ScopedCatalogPanel) ────────
 
@@ -266,6 +221,13 @@ function CatalogEntryForm({
     ? allParentEntries
     : [...scopeParentEntries, ...systemParentEntries.filter(e => !scopeParentEntries.some(s => s.id === e.id))];
   const baseEntry = parentEntries.find(e => e.id === baseRefId);
+  // Кандидаты базы для общего пикера (issue #73, шаг 2): записи родительского типа по скопам.
+  const baseCandidates: BaseCandidate[] = parentEntries
+    .map(e => ({
+      kind: 'catalog' as const, id: e.id, name: e.displayName, typeId: e.compositeTypeId,
+      tier: SCOPE_TIER[e.scope], scopeLabel: SCOPE_LABELS[e.scope], dist: 0,
+    }))
+    .sort((a, b) => a.tier - b.tier || a.name.localeCompare(b.name, 'ru'));
 
   // Наборы данных: биндинги существуют только у уже сохранённой записи (нужен id-владелец).
   const { data: bindings = [] } = useListDataSetBindings({ commonDataEntryId: entry?.id });
@@ -519,14 +481,11 @@ function CatalogEntryForm({
               Без базового экземпляра все {effectiveFields.length} полей заполняются вручную.
             </p>
           )}
-          <CatalogBaseEntryPicker
+          <BaseCandidatePicker
             open={basePickerOpen}
             onOpenChange={setBasePickerOpen}
-            parentType={parentType}
-            setId={setId}
-            scope={scope}
-            scopeId={scopeId}
-            onSelect={e => setValue('_baseRef', e.id)}
+            candidates={baseCandidates}
+            onSelect={c => setValue('_baseRef', c.id)}
           />
         </div>
       )}
